+++ default.nix
@@ -13,6 +13,7 @@ let
     interface = import ./lib/interface.nix;
     transport = import ./lib/transport.nix;
     hostname = import ./lib/hostname.nix;
+    domain = import ./lib/domain.nix;
     registry = import ./lib/registry.nix;
   };
 in